requirements.txt
Pillow==9.0.0
selenium==4.1.0
webdriver-manager==3.5.2
Flask